COM-1754 - get learner from company only in client interface
